\name{matrixY}
\alias{matrixY}
\docType{data}
\title{The treatment response matrix}
\description{A numeric matrix of dimension G by J, where G is the
total number of probesets measured by the microarray platform
and J is the total number of treatments. Each entry of matrixY
is the ratio of the probeset expression after and before treatment.}
\usage{data(matrixY)}
\format{
  The format is:
 num [1:50, 1:10] 0.6691 1.9856 -2.6227 0.0386 0.4526 ...
}
\examples{
data(matrixY)
}
\keyword{datasets}
